Added a Level system
Added a level system so now there are 10 blocks that are drawn from a
level1 class in a levels package.

I also reduced the gravity by 10 and reduced the angle that the platform
can roatate by a bit to make it not as hard.

Now with the different sized shapes textures are funky